+++ src/globals.h
@@ -1,0 +1,9 @@
+
+//use this file for global variables shared across multiple cpp files
+#pragma once 
+#include <Adafruit_BNO055.h>
+extern Adafruit_BNO055 bno;
+extern int chipSelect;
+extern int ledPin;
+extern int state;
+extern int flag;
